@@ -1,18 +1,40 @@
 [tool.black]
-target-version = ["py310"]
-
-[tool.isort]
-atomic = true
-combine_as_imports = true
-default_section = "THIRDPARTY"
-known_local_folder = [
-    "pacts",
-]
-known_third_party = [
-    "django",
-]
-profile = "black"
+target-version = ["py311"]
 
 [tool.mypy]
-python_version = "3.10"
+python_version = "3.11"
 ignore_missing_imports = true
+
+[tool.ruff]
+ extend-exclude = [
+     "__pycache__",
+     ".vscode*",
+ ]
+ extend-ignore = [
+    "D203",
+    "D212",
+    "D213",
+    "D214",
+    "D215",
+    "D404",
+    "D405",
+    "D406",
+    "D407",
+    "D408",
+    "D409",
+    "D410",
+    "D411",
+    "D413",
+    "D415",
+    "D416",
+    "D417",
+    "E501",
+ ]
+ select = ["B", "C", "D", "E", "F", "I", "W", "B9"]
+ target-version = "py311"
+
+[tool.ruff.isort]
+combine-as-imports = true
+known-third-party = [
+    "bootstrap",
+]
