@@ -1,7 +1,5 @@
 """Web project initialization CLI constants."""
 
-from typing import Dict
-
 # Stacks
 
 # BEWARE: stack names must be suitable for inclusion in Vault paths
@@ -39,7 +37,7 @@ DEV_ENV_NAME = "development"
 
 DEV_ENV_SLUG = "dev"
 
-DEV_ENV_STACK_CHOICES: Dict[str, str] = {
+DEV_ENV_STACK_CHOICES: dict[str, str] = {
     "1": MAIN_STACK_SLUG,
 }
 
@@ -47,7 +45,7 @@ STAGE_ENV_NAME = "staging"
 
 STAGE_ENV_SLUG = "stage"
 
-STAGE_ENV_STACK_CHOICES: Dict[str, str] = {
+STAGE_ENV_STACK_CHOICES: dict[str, str] = {
     "1": MAIN_STACK_SLUG,
     "2": DEV_STACK_SLUG,
 }
@@ -56,7 +54,7 @@ PROD_ENV_NAME = "production"
 
 PROD_ENV_SLUG = "prod"
 
-PROD_ENV_STACK_CHOICES: Dict[str, str] = {}
+PROD_ENV_STACK_CHOICES: dict[str, str] = {}
 
 # Env vars
 
